+++ lib/alg/dijkstra-all.ts
@@ -1,0 +1,8 @@
+import type { NodeKey } from '../graph';
+import type { DistanceMap, WeightFunc, EdgeFunc } from './types';
+import dijkstra from './dijkstra';
+import type Graph from '../graph';
+
+export default function dijkstraAll<K extends NodeKey>(g: Graph<K>, weightFunc?: WeightFunc<K>, edgeFunc?: EdgeFunc<K>): Map<K, DistanceMap<K>> {
+  return new Map(g.nodes().map(v => [v, dijkstra(g, v, weightFunc, edgeFunc)]));
+}